settings import glob + updates
