+++ src/modules/settings/components/Settings.jsx
@@ -1,15 +1,40 @@
 import React from 'react';
-// eslint-disable-next-line import/extensions, import/no-unresolved
-import './settings/*';
 import Panel from 'rsuite/Panel';
 import {Icon} from '@rsuite/icons';
 import InputGroup from 'rsuite/InputGroup';
 import AutoComplete from 'rsuite/AutoComplete';
 import * as faSearch from '@fortawesome/free-solid-svg-icons/faSearch';
-import {Components} from './Store.jsx';
 import FontAwesomeSvgIcon from '../../../common/components/FontAwesomeSvgIcon.jsx';
+import {loadModuleForPlatforms} from '../../../utils/modules.js';
+import {PlatformTypes} from '../../../constants.js';
 
-const settings = Object.values(Components).sort((a, b) => a.name.localeCompare(b.name));
+let settings = [];
+
+(async () => {
+  await loadModuleForPlatforms(
+    [
+      PlatformTypes.TWITCH,
+      async () => {
+        // eslint-disable-next-line import/no-unresolved
+        await import('./settings/global/*.jsx');
+        // eslint-disable-next-line import/no-unresolved
+        await import('./settings/twitch/*.jsx');
+      },
+    ],
+    [
+      PlatformTypes.YOUTUBE,
+      async () => {
+        // eslint-disable-next-line import/no-unresolved
+        await import('./settings/global/*.jsx');
+        // eslint-disable-next-line import/no-unresolved
+        await import('./settings/youtube/*.jsx');
+      },
+    ]
+  );
+
+  const {Components} = await import('./Store.jsx');
+  settings = Object.values(Components).sort((a, b) => a.name.localeCompare(b.name));
+})();
 
 export function Settings({search, category}) {
   const searchedSettings =
