--- src/modules/settings/components/settings/EmoteMenu.jsx
@@ -1,43 +1,0 @@
-import React from 'react';
-import Panel from 'rsuite/Panel';
-import Toggle from 'rsuite/Toggle';
-import {registerComponent} from '../Store.jsx';
-import {CategoryTypes, SettingIds} from '../../../../constants.js';
-import styles from '../../styles/header.module.css';
-import useStorageState from '../../../../common/hooks/StorageState.jsx';
-
-function EmoteMenu() {
-  const [emoteMenuValue, setEmoteMenuValue] = useStorageState(SettingIds.EMOTE_MENU);
-
-  return (
-    <Panel header="Emote Menu">
-      <div className={styles.toggle}>
-        <p>Enables a more advanced emote menu for Twitch</p>
-        <Toggle checked={emoteMenuValue} onChange={(state) => setEmoteMenuValue(state)} />
-      </div>
-    </Panel>
-  );
-}
-
-loadModuleForPlatforms(
-  [
-    PlatformTypes.TWITCH,
-    () =>
-      registerComponent(EmoteMenu, {
-        settingId: SettingIds.EMOTE_MENU,
-        name: 'Emote Menu',
-        category: CategoryTypes.CHAT,
-        keywords: ['twitch', 'emotes', 'popup'],
-      }),
-  ],
-  [
-    PlatformTypes.YOUTUBE,
-    () =>
-      registerComponent(EmoteMenu, {
-        settingId: SettingIds.EMOTE_MENU,
-        name: 'Emote Menu',
-        category: CategoryTypes.CHAT,
-        keywords: ['twitch', 'emotes', 'popup'],
-      }),
-  ]
-);
