--- src/modules/settings/components/settings/TabCompletionEmotePriority.jsx
@@ -1,31 +1,0 @@
-import React from 'react';
-import Panel from 'rsuite/Panel';
-import Toggle from 'rsuite/Toggle';
-import {registerComponent} from '../Store.jsx';
-import {CategoryTypes, SettingIds} from '../../../../constants.js';
-import styles from '../../styles/header.module.css';
-import useStorageState from '../../../../common/hooks/StorageState.jsx';
-
-function TabCompletionEmotePriority() {
-  const [value, setValue] = useStorageState(SettingIds.TAB_COMPLETION_EMOTE_PRIORITY);
-
-  return (
-    <Panel header="Tab Completion Emote Priority">
-      <div className={styles.toggle}>
-        <p>Prioritizes emotes over usernames when using tab completion</p>
-        <Toggle checked={value} onChange={(state) => setValue(state)} />
-      </div>
-    </Panel>
-  );
-}
-
-loadModuleForPlatforms([
-  PlatformTypes.TWITCH,
-  () =>
-    registerComponent(TabCompletionEmotePriority, {
-      settingId: SettingIds.TAB_COMPLETION_EMOTE_PRIORITY,
-      name: 'Tab Completion Emote Priority',
-      category: CategoryTypes.CHAT,
-      keywords: ['tab', 'completion', 'emote', 'priority'],
-    }),
-]);
