+++ src/modules/settings/components/settings/twitch/ChannelPoints.jsx
@@ -2,11 +2,11 @@ import React from 'react';
 import Panel from 'rsuite/Panel';
 import Checkbox from 'rsuite/Checkbox';
 import CheckboxGroup from 'rsuite/CheckboxGroup';
-import {registerComponent} from '../Store.jsx';
-import {SettingIds, CategoryTypes, ChannelPointsFlags} from '../../../../constants.js';
-import styles from '../../styles/header.module.css';
-import {hasFlag} from '../../../../utils/flags.js';
-import useStorageState from '../../../../common/hooks/StorageState.jsx';
+import {registerComponent} from '../../Store.jsx';
+import {SettingIds, CategoryTypes, ChannelPointsFlags} from '../../../../../constants.js';
+import styles from '../../../styles/header.module.css';
+import {hasFlag} from '../../../../../utils/flags.js';
+import useStorageState from '../../../../../common/hooks/StorageState.jsx';
 
 function ChannelPointsModule() {
   const [channelPoints, setChannelPoints] = useStorageState(SettingIds.CHANNEL_POINTS);
@@ -36,13 +36,9 @@ function ChannelPointsModule() {
   );
 }
 
-loadModuleForPlatforms([
-  PlatformTypes.TWITCH,
-  () =>
-    registerComponent(ChannelPointsModule, {
-      settingId: SettingIds.CHANNEL_POINTS,
-      name: 'Channel Points',
-      category: CategoryTypes.CHAT,
-      keywords: ['channel', 'points', 'auto', 'claim'],
-    }),
-]);
+registerComponent(ChannelPointsModule, {
+  settingId: SettingIds.CHANNEL_POINTS,
+  name: 'Channel Points',
+  category: CategoryTypes.CHAT,
+  keywords: ['channel', 'points', 'auto', 'claim'],
+});
