+++ src/modules/settings/components/settings/twitch/Chat.jsx
@@ -2,11 +2,11 @@ import React from 'react';
 import Panel from 'rsuite/Panel';
 import Checkbox from 'rsuite/Checkbox';
 import CheckboxGroup from 'rsuite/CheckboxGroup';
-import {registerComponent} from '../Store.jsx';
-import {CategoryTypes, SettingIds, ChatFlags} from '../../../../constants.js';
-import styles from '../../styles/header.module.css';
-import {hasFlag} from '../../../../utils/flags.js';
-import useStorageState from '../../../../common/hooks/StorageState.jsx';
+import {registerComponent} from '../../Store.jsx';
+import {CategoryTypes, SettingIds, ChatFlags} from '../../../../../constants.js';
+import styles from '../../../styles/header.module.css';
+import {hasFlag} from '../../../../../utils/flags.js';
+import useStorageState from '../../../../../common/hooks/StorageState.jsx';
 
 function ChatModule() {
   const [chat, setChat] = useStorageState(SettingIds.CHAT);
@@ -52,13 +52,9 @@ function ChatModule() {
   );
 }
 
-loadModuleForPlatforms([
-  PlatformTypes.TWITCH,
-  () =>
-    registerComponent(ChatModule, {
-      settingId: SettingIds.CHAT,
-      name: 'Chat',
-      category: CategoryTypes.CHAT,
-      keywords: ['bits', 'highlights', 'community', 'chat', 'replies', 'clips', 'subs', 'subscriptions'],
-    }),
-]);
+registerComponent(ChatModule, {
+  settingId: SettingIds.CHAT,
+  name: 'Chat',
+  category: CategoryTypes.CHAT,
+  keywords: ['bits', 'highlights', 'community', 'chat', 'replies', 'clips', 'subs', 'subscriptions'],
+});
