+++ src/modules/settings/components/settings/twitch/DeletedMessages.jsx
@@ -3,10 +3,10 @@ import Panel from 'rsuite/Panel';
 import Radio from 'rsuite/Radio';
 import RadioGroup from 'rsuite/RadioGroup';
 import FormGroup from 'rsuite/FormGroup';
-import {registerComponent} from '../Store.jsx';
-import {CategoryTypes, SettingIds, DeletedMessageTypes} from '../../../../constants.js';
-import styles from '../../styles/header.module.css';
-import useStorageState from '../../../../common/hooks/StorageState.jsx';
+import {registerComponent} from '../../Store.jsx';
+import {CategoryTypes, SettingIds, DeletedMessageTypes} from '../../../../../constants.js';
+import styles from '../../../styles/header.module.css';
+import useStorageState from '../../../../../common/hooks/StorageState.jsx';
 
 function DeletedMessagesModule() {
   const [value, setValue] = useStorageState(SettingIds.DELETED_MESSAGES);
@@ -41,14 +41,9 @@ function DeletedMessagesModule() {
     </Panel>
   );
 }
-
-loadModuleForPlatforms([
-  PlatformTypes.TWITCH,
-  () =>
-    registerComponent(DeletedMessagesModule, {
-      settingId: SettingIds.DELETED_MESSAGES,
-      name: 'Deleted Messages',
-      category: CategoryTypes.CHAT,
-      keywords: ['messages', 'deleted'],
-    }),
-]);
+registerComponent(DeletedMessagesModule, {
+  settingId: SettingIds.DELETED_MESSAGES,
+  name: 'Deleted Messages',
+  category: CategoryTypes.CHAT,
+  keywords: ['messages', 'deleted'],
+});
