+++ src/modules/settings/components/settings/twitch/PinnedHighlights.jsx
@@ -1,0 +1,27 @@
+import React from 'react';
+import Panel from 'rsuite/Panel';
+import Toggle from 'rsuite/Toggle';
+import {registerComponent} from '../../Store.jsx';
+import {CategoryTypes, SettingIds} from '../../../../../constants.js';
+import styles from '../../../styles/header.module.css';
+import useStorageState from '../../../../../common/hooks/StorageState.jsx';
+
+function PinnedHighlights() {
+  const [value, setValue] = useStorageState(SettingIds.PINNED_HIGHLIGHTS);
+
+  return (
+    <Panel header="Pinned Highlights">
+      <div className={styles.toggle}>
+        <p>Pins your last ten highlighted messages above chat</p>
+        <Toggle checked={value} onChange={(state) => setValue(state)} />
+      </div>
+    </Panel>
+  );
+}
+
+registerComponent(PinnedHighlights, {
+  settingId: SettingIds.PINNED_HIGHLIGHTS,
+  name: 'Pinned Highlights',
+  category: CategoryTypes.CHAT,
+  keywords: ['pinned', 'highlights'],
+});
