+++ src/modules/settings/components/settings/youtube/YouTube.jsx
@@ -1,10 +1,10 @@
 import React, {useEffect} from 'react';
 import Panel from 'rsuite/Panel';
 import Toggle from 'rsuite/Toggle';
-import {registerComponent} from '../Store.jsx';
-import {CategoryTypes} from '../../../../constants.js';
-import styles from '../../styles/header.module.css';
-import extension from '../../../../utils/extension.js';
+import {registerComponent} from '../../Store.jsx';
+import {CategoryTypes} from '../../../../../constants.js';
+import styles from '../../../styles/header.module.css';
+import extension from '../../../../../utils/extension.js';
 
 const EXTENSION_URL = new URL(extension.url(''));
 const HAS_LOCAL_EXTENSION_DATA = EXTENSION_URL.protocol.includes('extension');
