+++ src/modules/settings/styles/_rsuite.less
@@ -19,6 +19,7 @@
 .@{ns}modal {
   width: 800px;
   height: 500px;
+
   color: var(--bttv-rs-text-primary);
   font-size: 1.3rem;
   line-height: 1.5;
@@ -41,6 +42,10 @@
   *::after {
     box-sizing: border-box;
   }
+}
+
+.@{ns}auto-complete-item {
+  width: unset;
 }
 
 .@{ns}modal {
