+++ src/utils/window.js
@@ -47,3 +47,7 @@ export function isPopout() {
     return true;
   }
 }
+
+export function isExternalWindow() {
+  return window.location.pathname.endsWith('/chat') || window.location.pathname.endsWith('/live_chat');
+}
